requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
WTForms==3.1.1
plotly
requests
pandas
google-generativeai==0.3.2
gunicorn
